Add inventory & refine level select UI
Added a font, nineslice background, and got the inventory all working. Collected tems are defined in the global script and we can change that item's count. Format for an item array is shown in the global script. Level select UI also has been updated a bit, but still lacks functionality.